+++ qlty-config/src/workspace.rs
@@ -149,6 +149,11 @@ impl Workspace {
                 return Some(current);
             }
 
+            if let Ok(_) = std::env::var("QLTY_COVERAGE_TESTING_WITHOUT_GIT") {
+                // If we're in testing mode, we might not have a .git directory.
+                return None;
+            }
+
             if !current.pop() {
                 return None;
             }